Move the PasswordChangingPostgresDataSource and uncomment code
